# backend/internal/httpd/apispec/openapi.yaml
openapi: 3.1.0
info:
  description: Loopback-only HTTP surface served by the Go daemon. Generated from
    Go (code-first) — do not edit by hand; run `go generate ./...`.
  title: Thanos HTTP daemon
  version: 0.1.0-route-shell
servers:
- description: Local daemon (loopback only)
  url: http://127.0.0.1:3001
paths:
  /api/v1/agents:
    get:
      operationId: listAgents
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAgentsResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Return cached supported and locally installed agent adapters
      tags:
      - agents
  /api/v1/agents/{agent}/probe:
    post:
      operationId: probeAgent
      parameters:
      - description: Agent adapter identifier.
        in: path
        name: agent
        required: true
        schema:
          description: Agent adapter identifier.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProbeAgentResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Run a fresh local readiness probe for one agent adapter
      tags:
      - agents
  /api/v1/agents/refresh:
    post:
      operationId: refreshAgents
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAgentsResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Refresh the cached local agent adapter catalog
      tags:
      - agents
...
  /api/v1/notifications:
    get:
      operationId: listNotifications
      parameters:
      - description: Notification status filter. V1 supports only unread.
        in: query
        name: status
        schema:
          description: Notification status filter. V1 supports only unread.
          enum:
          - unread
          type: string
      - description: Maximum notifications to return. Defaults to 50; capped at 100.
        in: query
        name: limit
        schema:
          description: Maximum notifications to return. Defaults to 50; capped at
            100.
          maximum: 100
          minimum: 1
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListNotificationsResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: List unread notifications
      tags:
      - notifications
  /api/v1/notifications/{id}:
    patch:
      operationId: markNotificationRead
      parameters:
      - description: Notification identifier.
        in: path
        name: id
        required: true
        schema:
          description: Notification identifier.
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarkNotificationReadRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationEnvelope'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Mark a notification read
      tags:
      - notifications
  /api/v1/notifications/read-all:
    post:
      operationId: markAllNotificationsRead
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarkAllNotificationsReadResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Mark all unread notifications read
      tags:
      - notifications
  /api/v1/notifications/stream:
    get:
      operationId: streamNotifications
      parameters:
      - description: Optional project id filter for live notifications.
        in: query
        name: projectId
        schema:
          description: Optional project id filter for live notifications.
          type: string
      responses:
        "200":
          content:
            text/event-stream:
              schema:
                type: string
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Stream created notifications
      tags:
      - notifications
...
components:
  schemas:
    APIError:
      properties:
        code:
          type: string
        details:
          additionalProperties: {}
          type: object
        error:
          type: string
        message:
          type: string
        requestId:
          type: string
      required:
      - error
      - code
      - message
      type: object
...
    AgentInfo:
      properties:
        authStatus:
          description: Advisory local auth probe result. authorized means a recent
            local probe passed; spawn remains the authoritative validation point.
          enum:
          - authorized
          - unauthorized
          - unknown
          type: string
        id:
          type: string
        label:
          type: string
      required:
      - id
      - label
      type: object
...
    ListAgentsResponse:
      properties:
        authorized:
          description: Compatibility list of installed agents whose local auth probe
            recently returned authorized. Advisory and stale-prone; spawn may still
            fail.
          items:
            $ref: '#/components/schemas/AgentInfo'
          type: array
        installed:
          description: Agents whose binary resolved during the latest best-effort
            local catalog probe.
          items:
            $ref: '#/components/schemas/AgentInfo'
          type: array
        supported:
          description: Agents supported by this daemon build.
          items:
            $ref: '#/components/schemas/AgentInfo'
          type: array
      required:
      - supported
      - installed
      - authorized
      type: object
    ListNotificationsResponse:
      properties:
        notifications:
          items:
            $ref: '#/components/schemas/NotificationResponse'
          type: array
      required:
      - notifications
      type: object
...
    MarkAllNotificationsReadResponse:
      properties:
        notifications:
          items:
            $ref: '#/components/schemas/NotificationResponse'
          type: array
      required:
      - notifications
      type: object
    MarkNotificationReadRequest:
      properties:
        status:
          description: V1 supports only marking an unread notification read.
          enum:
          - read
          type: string
      required:
      - status
      type: object
...
    NotificationEnvelope:
      properties:
        notification:
          $ref: '#/components/schemas/NotificationResponse'
      required:
      - notification
      type: object
    NotificationResponse:
      properties:
        body:
          type: string
        createdAt:
          format: date-time
          type: string
        id:
          type: string
        prUrl:
          type: string
        projectId:
          type: string
        sessionId:
          type: string
        status:
          enum:
          - unread
          - read
          type: string
        target:
          $ref: '#/components/schemas/NotificationTarget'
        title:
          type: string
        type:
          enum:
          - needs_input
          - ready_to_merge
          - pr_merged
          - pr_closed_unmerged
          type: string
      required:
      - id
      - sessionId
      - projectId
      - prUrl
      - type
      - title
      - body
      - status
      - createdAt
      - target
      type: object
    NotificationTarget:
      properties:
        kind:
          enum:
          - session
          - pr
          type: string
        prUrl:
          type: string
        sessionId:
          type: string
      required:
      - kind
      - sessionId
      type: object
...
    ProbeAgentResponse:
      properties:
        agent:
          $ref: '#/components/schemas/AgentInfo'
        installed:
          type: boolean
        supported:
          type: boolean
      required:
      - agent
      - supported
      - installed
      type: object
...